Add option to normalize input when calling get_input

+++ rlgammon/environment/utils/normalize_input.py
@@ -76,7 +76,7 @@ cell_stats = [
 
 def normalize_input(arr: Input, stats: list[tuple[float, float]]) -> Input:
     """Normalize the observations."""
-    standardized = np.empty_like(arr, dtype=float)
+    standardized = np.empty_like(arr, dtype=np.float16)
     for i, (mean, std) in enumerate(stats):
         standardized[i] = (arr[i] - mean) / std
     return standardized
